packages/evremixes/pyproject.toml
[project]
name = "evremixes"
version = "0.1.0"
description = "Evanescence Remix Downloader"
authors = [{ name = "Danny Stewart", email = "danny@stewart.cc" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
urls = { Homepage = "https://music.dannystewart.com/evanescence/", Repository = "https://github.com/dannystewart/dsbase/packages/evremixes" }
dependencies = [
    "dsbase",
    "halo>=0.0.31,<0.0.32",
    "inquirer>=3.4.0,<4.0.0",
    "mutagen>=1.47.0,<2.0.0",
    "pillow>=11.1.0,<12.0.0",
    "requests>=2.32.3,<3.0.0",
    "termcolor>=2.5.0,<3.0.0",
    "types-requests>=2.32.0.20250306,<3.0.0.0",
]

[tool.uv.sources]
dsbase = { workspace = true }

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
evremixes = "evremixes.main:main"
